+++ lib/network/auth_service.dart
@@ -31,6 +31,10 @@ class AuthService {
 
   final Logger log = sl.get<Logger>();
 
+  // curl --request GET 'https://auth.perish.co/seed-backup/:id'
+
+  // curl --request POST 'https://auth.perish.co/seed-backup/:id'
+
   Future<String?> getEncryptedSeed(String identifier) async {
     try {
       final http.Response resp = await http.get(
@@ -65,5 +69,37 @@ class AuthService {
     }
 
     return true;
+  }
+
+  Future<bool> entryExists(String identifier) async {
+    try {
+      final http.Response resp = await http.get(
+        Uri.parse("$AUTH_SERVER/seed-exists/$identifier"),
+        headers: <String, String>{"Accept": "application/json"},
+      );
+      if (resp.body.contains("not_found")) {
+        return false;
+      }
+      return true;
+    } catch (e) {
+      log.e(e);
+      return true;
+    }
+  }
+
+  Future<bool> deleteEncryptedSeed(String fullIdentifier) async {
+    try {
+      final http.Response resp = await http.get(
+        Uri.parse("$AUTH_SERVER/delete-seed/$fullIdentifier"),
+        headers: {"Accept": "application/json"},
+      );
+      if (resp.statusCode == 200) {
+        return true;
+      }
+      return false;
+    } catch (e) {
+      log.e(e);
+      return false;
+    }
   }
 }